# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salus API",
        "description": "API documentation for the Salus health assistant application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Application server"
        }
    ],
    "paths": {
...
        "/api/ai/history": {
            "get": {
                "tags": [
                    "AI"
                ],
                "summary": "Get AI advice history",
                "operationId": "listHealthAdviceHistory",
                "responses": {
                    "200": {
                        "description": "AI advice history retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "AI advice history retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "advice": {
                                                        "type": "string",
                                                        "example": "Stay hydrated, rest well, and monitor your symptoms."
                                                    },
                                                    "symptoms": {
                                                        "type": "string",
                                                        "example": "Headache, Fatigue"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-03-26T11:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-03-26T11:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use a Sanctum bearer token. Example: Bearer 1|abcdef...",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}